Add useUser hook to simplify requesting for authenticated user's profile
