+++ src/Components/Account.js
@@ -1,29 +1,13 @@
-import { useEffect, useContext } from "react";
-import UserContext from "../Context/UserProvider.js";
 import { useNavigate } from "react-router-dom";
 import useAuth from "../Hooks/useAuth.js";
+import useUser from "../Hooks/useUser.js";
 import useSignOut from "../Hooks/useSignOut.js";
-import useAxiosPrivate from "../Hooks/useAxiosPrivate.js";
 
 export default function Account() {
   const navigate = useNavigate();
   const { auth } = useAuth();
-  const axiosPrivate = useAxiosPrivate();
-  const { user, setUser } = useContext(UserContext);
+  const { user } = useUser();
   const signOut = useSignOut();
-
-  useEffect(() => {
-    if (auth?.token) {
-      loadUserProfile();
-    }
-  }, [auth]);
-
-  const loadUserProfile = async () => {
-    const profile = await axiosPrivate.get("/user-profile", {
-      withCredentials: true,
-    });
-    setUser(profile.data);
-  };
 
   const handleSignOut = async () => {
     await signOut();
@@ -33,7 +17,7 @@ export default function Account() {
   return (
     <div className="App">
       <header className="App-header">
-        {user.username ? (
+        {user?.username ? (
           <h1 className="large bold">{`Welcome, ${user?.username}!`}</h1>
         ) : (
           <h1 className="large bold">Welcome!</h1>
